Move ImGui Json helpers to Editor
